requirements.txt
altair==6.0.0
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.62.1
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
multitasking==0.0.12
narwhals==2.18.1
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
parso==0.8.6
patsy==1.0.2
peewee==4.0.0
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.47
stack-data==0.6.3
statsmodels==0.14.6
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
websockets==16.0
yfinance==1.2.0
